+++ src/app/locales/en/components/content-clusters-members.js
@@ -5,6 +5,8 @@ export default {
   directUsers: 'Direct users',
   effectiveGroups: 'Effective groups',
   effectiveUsers: 'Effective users',
+  noConnectedUserWarningHeader: 'This cluster does not have any linked Onezone user',
+  noConnectedUserWarning: 'It is strongly recommended to link at least one user to ensure maximum control in case of any failures and gain convenient access via unified Onezone interface. To link your user, please generate a token below, log in to Onezone and join this cluster in the Clusters menu.',
   openInOnezone: 'Open in Onezone',
   inviteUserUsingToken: 'Invite user using a token',
   generateAnotherToken: 'Generate another token',