+++ src/app/locales/en/components/sidebar-clusters.js
@@ -2,5 +2,6 @@ export default {
   secondLevelItems: {
     dnsWarning: 'Some problems with your DNS configuration have been detected',
     webCertWarning: 'Some problems with your web certificate have been detected',
+    noConnectedUserWarning: 'This cluster does not have any linked Onezone user',
   },
 };